Tidal: Fix missing countryCode on pin invoke(#7318)
Previously the 'UserId', 'CountryCode' and 'Username' values were
being stored in a struct 'UserInfo' from a  std::map keyed on the
'TokenId' of the associated OAuth token.

Silly Ewan was wrapping this 'TokenId' value in a Brn(), whose
memory was being freed from underneath it.

Fix was to store 'UserInfo' structs with a copy of the 'TokenId'
for easy comparisons. Internally they are now stored in a vector of
size 'kMaximumSupportedTokens' and searched linearly on demand.