TIDAL: Fix bug where tokens could fail to refresh
Internally a cache is kept of the currently logged in userId and
CountryCode in order to allow streaming.

Previously, when a token refreshed, it would only ever use a new slot in
this cache. Once the cache was full, then claim to have failed to
refresh.

Fix ensures we replace the already existing entry for the given token
before we populate a new one.

Have also updated logging around this case so it can be better
identified.