[TIDAL] Prevent reading from an unconnected socket
Commits 47ae2153 and d3280fa4 has introduced a logging pathway for
streaming errors in TIDAL code. Unfortunately both HTTP and Socket
errors are communicated back to callers in the same manner.

In the case a socket failed to connect, we'd then attempt to read from
buffers that were either in an invalid state or referencing a previous
network request. This usually manifested itself as an assert or a seg
fault.

See #7721 for more details.