Json: ParserArray::TryNext___()
Instead of throwing JsonArrayEnumerationComplete and using it control
execution flow, this patch adds a TryNext___() variant that returns
True/False indicating if the parse was a success.

Making use of this should speed up JSON parsing performance as we are
doing less work each time we find there are no more items to parse.