ohMetadata: Split out DIDL XML writing to seperate helper class
There are a number of places where we construct DIDL metadata for items
throughout our codebase, most of which re-implement the functionality in
a variety of different ways.

This patch aims to start condensing these implementations down by having
a common set of DIDL XML components for use. To start with, the
ohMetadata class for writing metadata has been split into 2. The
underlying DIDL XMl writing functions are now found in WriterDIDLXml.

This can then be used by later components that provide a more pleasant &
flexible interface, instead of 'ohMetadata'