TIDAL: Convert pins to use standard DIDLWriter class
This patch removes the custom DIDL-Lite writing code from inside TIDAL
pins. By doing so, we are able to remove a large amount of duplicated
code.

We also can take advantage of the restructuring an do less parsing work,
particularly when adding artworks. Previously, for each type of artwork,
we had to re-parse the item ID, do some formatting for each artwork
resolution. Now, we do that work once and instead do some buffer
operations. In theory this should be a lot quicker than before.

Additionally, we now report the type of exception if one is thrown when
turning TIDAL metadata into DIDL-Lite.