Pins: Don't check total number of items when fetching track
Pins that were fetching a single track would make 2 network requests.
The first to work out how many items in total to download, then a
request to download the item.

In the case of a single track, this first request is wasteful as it
would contain the metadata for the item. Instead, we shortcut to assume
a track is a single item and save a network request.