MPD: Initial helper classes for parsing MPEG-DASH manifest files
A bunch of helper classes for parsing MPD documents. They are currently a little
WIP as they are used in upstream projects.

From what I can gather in the spec, every service is likely to format their DASH
manifests differently and so we'll probably require custom parsing classes for
each service...