Pipeline: Add basic support for MPEG-DASH parsing
This patch adds a hook into the pipeline for MPEG-DASH processing. This
exists as a 'ContentProcessor' instance that recognises the 'MPD'
MIME-Type.

Clients should then register implementations of MPD parses with the
pipeline. Each registered instance will be queried to check if we they
can handle the incoming MPD.

NOTE: This structure is subject to change