Product: Don't notify Uri/Attribute changes if nothing has changed
Product watches for network adapter changes and events out to observers
once this has been done.

However, we can be notified of network adapter changes which don't
result in changes in any of the values this class or it's observers
truely care about.

Most notably, this was causing Core1 devices to do lots more work,
ultimately resulting in the device eventing a BYEBYE/ALIVE due to a
property changed a number of seconds after booting.

In this patch we capture the old value(s) before they are changed. They
are them compared with the new values and ONLY if there has been a
change do observers get notified.

See #8271 for more details