Qobuz: Reduce number of allocated buffers for parsing JSON in Pins
Removed the need for a number of intermediate buffers to store parent
object metadata.

Firstly - we know we have the full response in memory so for any of the
values that we need to unescape, we can do that in place as the result
will always be the same size (or smaller) overall.

Secondly - As the parent metadata exists in memory for the duration of
the track generation for that repsonse chunk we don't need to allocate
and copy the values elsewhere. Instead we use raw Brn values to point
into the response buffer

Thirdly - Use of std::function/lambda for a shared JSON unescaping
function is unnecessary.