MPEG: Ensure tracks with empty TRUN sample table can be played
If the TRUN box sample table is empty, we can still play those tracks
assuming that the TFHD contains a 'default_sample_size' header. In the
case of Opus tracks, they have a fixed sample size and so only specify
the default size and no sample table to save space in the MPEG stream.