MPEG: Changes to Mpeg4Container Metadata handling
Previously assumed that the metadata would only be required at most
once per-stream. But in some cases where you have a fragmented stream,
certain codecs might require some metadata on a per-fragment level.

This patch adjusts the logic used to emit the required metadata depending
on the codec and state of the stream.