Codec/Build: Fix Opus building test executables and unecessary tests
Opus was including files that were tests in their own right. This meant
that some of our test executables were attempting to run using the Opus
main() method instead of their own resulting in a bunch of errors.

This has been addressed and have also removed a number of other files from
the Opus build that deal with features we don't use.