MPEG: Fix crash when dealing with large sample sizes
Some of the DRM protected files we encounter have samples larger than
the 9216 bytes we can fit into a MsgAudioEncoded instance. This
previously caused us to assert and crash trying to split the message on
0 bytes.

This patch addresses this by buffering more audio data and only emitting
a full MsgAudioEncoded when required. Furthermore, if we've not
increased buffer sizes enough, we should no longer crash and instead
stop the current track and move onto the next.