Pipeline: Call SetFlushing on AudioEncodedCache during TryStop
Previously, without this call, upon stopping or skipping MPEG tracks
would require us to continue reading the MPEG container via the cache
until we ran out of data (or another problem occured).

Normally this was fine, but with the addition of some DRM, this added
a large amount of overhead to this process - sometimes 5 seconds worth
- between tracks.

It was noticed that we call SetFlushing() on the AudioEncoded cache during
ContainerTryStop and TrySeek but not TryStop. By adding this, we no longer
get the long duration between tracks and we no longer attempt to read from
MPEG container after we've decided to stop playing the current track!