MPEG: Move IsFragment from SeekTable to Mpeg4Info
For some reason, which I don't fully understand, the addition of the is
'IsFragmented' flag to the MPEG SeekTable class was causing the FDk and ALAC
codecs to fail to play tracks - mostly with corrupted stream errors.

It's likely that this additional byte was causing off-by-one errors in the
parsing. Which really doesn't make much sense as all the read and write
callsites for that particular class accounted for this.

Originally it was deemed the 'IsFragmented' information was required only
for seeking, but it might be useful for other things in the future such
as supporting fragmented radio streams that'll require some codec changes.

Thus, the 'IsFragmented' flag has been moved from SeekTable -> Mpeg4Info.
I suspect this wont cause any problems for the codecs as the Mpeg4Info is
always read first, before any codec specific initialisation data.

Fixes #8794